Move Notifications module into Settings package
