+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/settings/notifications/NotificationsModule.kt
@@ -1,11 +1,10 @@
-package io.horizontalsystems.bankwallet.modules.notifications
+package io.horizontalsystems.bankwallet.modules.settings.notifications
 
 import android.content.Context
 import android.content.Intent
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import io.horizontalsystems.bankwallet.core.App
-import io.horizontalsystems.bankwallet.entities.Coin
 
 object NotificationsModule {
